hw2/requirements.txt
tqdm
tensorboard
torch
torchvision
pyyaml
numpy
